requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
boto3==1.34.69
botocore==1.34.69
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distro==1.8.0
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.3.3
greenlet==3.0.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
ngrok==0.11.0
openai==1.3.5
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
slackclient==2.9.4
sniffio==1.3.0
SQLAlchemy==2.0.21
tqdm==4.65.0
typing_extensions==4.8.0
urllib3==2.0.3
Werkzeug==2.3.5
yarl==1.9.2
